Refactor bot and NLP modules
