+++ app/nlp/__init__.py
@@ -1,5 +1,5 @@
 from .preprocessor import TextPreprocessor
-from .classifier import FeedbackClassifier
+from .classifier import TextClassifier
 from .embedder import TextEmbedder
 
-__all__ = ['TextPreprocessor', 'FeedbackClassifier', 'TextEmbedder'] 
+__all__ = ['TextPreprocessor', 'TextClassifier', 'TextEmbedder']
