+++ app/nlp/classifier.py
@@ -6,7 +6,7 @@ import torch
 
 logger = logging.getLogger(__name__)
 
-class FeedbackClassifier:
+class TextClassifier:
     def __init__(self, model_name: str = "facebook/bart-large-mnli", use_gpu: bool = False):
         """
         Initialize the feedback classifier.
